apps/api/pyproject.toml
[project]
name = "travel-pwa-api"
version = "0.1.0"
description = "FastAPI GraphQL backend for Travel PWA"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "strawberry-graphql[fastapi]>=0.260.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "python-dotenv>=1.0.1",
  "pyjwt[crypto]>=2.9.0",
  "httpx>=0.27.2",
  "boto3>=1.35.45",
  "pydantic-settings>=2.6.1"
]

[tool.uv]
dev-dependencies = [
  "ruff>=0.7.1"
]
